==> notebooks/01_data_cleaning.ipynb <==
# 01 - Data Cleaning

import os
import sys

module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

import pandas as pd
from src import data_cleaning

# Limitar la cantidad de filas cargadas (porque el dataset es demasiado grande, casi 1M de registros, y se cuelga)
MAX_ROWS = 10000
df = pd.read_csv('../data/raw/properati.csv', nrows=MAX_ROWS)
df_clean = data_cleaning.clean_data(df)
df_clean.to_csv('../data/processed/properati_clean.csv', index=False)
print(f'Saved processed sample ({len(df_clean)} rows).')
